Move FormEventTypeExtension into the Bolt\Form\Type namespace
